requirements.txt
python-dotenv>=1.0.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
fastapi>=0.104.0
uvicorn>=0.24.0
transformers>=4.35.0
torch>=2.1.0
accelerate>=0.24.0
httpx>=0.25.0
